refactor(capture): unify ffmpeg backends into shared recorder
Extract the common ffmpeg pipeline shared by the screencast and
screenshot backends into a new recorder module. Each backend now only
supplies its three genuine differences (sizing, source, label),
collapsing ~210 lines of duplicated logic.
